+++ fut/__init__.py
@@ -23,7 +23,7 @@ Basic usage:
 """
 
 __title__ = 'fut'
-__version__ = '0.3.0'
+__version__ = '0.3.1'
 __author__ = 'Piotr Staroszczyk'
 __author_email__ = 'piotr.staroszczyk@get24.org'
 __license__ = 'GNU GPL v3'